Draw into a buffered image, then draw it onto the screen.
With the previous implementation, the paint code only works with factors
of the numbers of bars (256, which means 256px, 512px, etc). Calculate
the smallest multiple which is larger than the panel's size (so 512 if
the panel is 500px wide), then create a transparent BufferedImage and
draw into that. This is equivalent to drawing into an off-screen buffer.
Once the bars have drawn, draw it into the panel (which will scale it
down to the window size, ex 512 -> 500).