Improvements to Rendering Code (#9)
* Use setContentPane() instead of setVisible

This prevents issues with resizing the window after returning from a
sort display

* Use the for(;;;) format for instead of for(X x:y)

Since we already need to increment an index value, this makes the code
more clear.

* Render the bars at different heights depending on window height.

* Create and dispose a new graphics when drawing.

Java's Graphics2D works similarly to OpenGL. create() is somewhat
similar to pushMatrix, dispose() is similar to popMatrix. Creating and
disposing graphics will ensure that you don't accidentally
transform/change colors of other Swing components.

* Draw into a buffered image, then draw it onto the screen.

With the previous implementation, the paint code only works with factors
of the numbers of bars (256, which means 256px, 512px, etc). Calculate
the smallest multiple which is larger than the panel's size (so 512 if
the panel is 500px wide), then create a transparent BufferedImage and
draw into that. This is equivalent to drawing into an off-screen buffer.
Once the bars have drawn, draw it into the panel (which will scale it
down to the window size, ex 512 -> 500).

* Allow more than 256 bars to render correctly.

This changes the way bars render: Previously, the height of bars was
calculated as (value * 2) / 720. This does not scale well past ~350 bars,
as they go off screen. Bar height on screen is now calculated as (value
/ maxValue) * (512 / 720) * panelHeight. This ensures that bars only
take up roughly the same amount of screen space as before, but now many
more bars will work without rendering off the screen.